# Conflicts:
#	AGI/AGI.h
#	Erebus/Erebus.vcxproj.filters
#	Erebus/GamePlay.cpp
#	Erebus/LuaAi.cpp
#	Erebus/LuaBinds.cpp
#	Erebus/Materials/blinn1.material
#	Erebus/Models/Goblin.model
#	Erebus/Scripts/AI/BasicEnemyAI.lua
#	Erebus/Scripts/AI/states.lua
#	Erebus/Scripts/Animation_Controllers/meleeGoblinController.lua
#	Erebus/Scripts/enemies.lua
#	Erebus/Scripts/level.lua
#	Erebus/Scripts/main.lua
#	Erebus/Scripts/player.lua
#	Erebus/source.cpp

+++ Erebus/Menu.cpp
@@ -15,6 +15,7 @@ Menu::Menu(Gear::GearEngine* inEngine, Importer::Assets* assets)
 	clientButton = sScreenImage(glm::vec2(350, 550), 350, 100);
 	clientButtonTex = assets->load<TextureAsset>("Textures/buttonConnect.png");
 
+
 }
 
 Menu::~Menu()
@@ -25,17 +26,17 @@ Menu::~Menu()
 enum GameState Menu::Update(Inputs* inputs)
 {
 
-	if (button.mousePick(inputs->getMousePos().x, inputs->getMousePos().y) && inputs->buttonReleasedThisFrame(0))
+	if (button.mousePick((float)inputs->getMousePos().x, (float)inputs->getMousePos().y) && inputs->buttonReleasedThisFrame(0))
 	{
 
 		return  GameplayState;
 	}
-	else if (hostButton.mousePick(inputs->getMousePos().x, inputs->getMousePos().y) && inputs->buttonReleasedThisFrame(0))
+	else if (hostButton.mousePick((float)inputs->getMousePos().x, (float)inputs->getMousePos().y) && inputs->buttonReleasedThisFrame(0))
 	{
 
 		return  HostGameplayState;
 	}
-	else if (clientButton.mousePick(inputs->getMousePos().x, inputs->getMousePos().y) && inputs->buttonReleasedThisFrame(0))
+	else if (clientButton.mousePick((float)inputs->getMousePos().x, (float)inputs->getMousePos().y) && inputs->buttonReleasedThisFrame(0))
 	{
 
 		return  ClientGameplayState;
